backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
tensorflow==2.15.0
pillow==10.1.0
python-dotenv==1.0.0
aiosqlite==0.19.0
